Fix decoder plugin not throwing the missing dependecy err msg issue (#224)
The expected behaviour for the optional decoder plugin is that if any of
the python dependencies is missing and the user attempts to use the
optional decoder, a runtime error message should be thrown. Adding
"ImportError" to catch the missing dependency issue and therefore
trigger the expected runtime error message.

---------

Signed-off-by: Melody Ren <melodyr@nvidia.com>